Updated the styling to be more compact
