+++ src/screens/SavedRecipes/SavedRecipesScreen.js
@@ -14,6 +14,7 @@ import {
 } from "../../utils/APICalls/SimplKitchen/userRecipes";
 import { useFocusEffect } from "@react-navigation/native";
 import MenuButton from "../../components/MenuButton/MenuButton";
+import styles from "./styles";
 
 const SavedRecipesScreen = ({ navigation }) => {
   const [savedRecipes, setSavedRecipes] = useState(null);
@@ -119,81 +120,6 @@ const SavedRecipesScreen = ({ navigation }) => {
       )}
     </View>
   );
-};
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    padding: 10,
-  },
-  noRecipesContainer: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-  },
-  noRecipesText: {
-    fontSize: 18,
-    fontWeight: "bold",
-    textAlign: "center",
-  },
-  recipeList: {
-    flex: 1,
-    width: "100%",
-  },
-  recipeContainer: {
-    borderWidth: 1,
-    borderColor: "#ccc",
-    borderRadius: 10,
-    padding: 10,
-    marginBottom: 10,
-    backgroundColor: "#fff",
-  },
-  recipeTitleContainer: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    alignItems: "center",
-    marginBottom: 10,
-  },
-  recipeTitle: {
-    fontSize: 18,
-    fontWeight: "bold",
-  },
-  deleteButton: {
-    backgroundColor: "#97DF99",
-    paddingHorizontal: 20,
-    paddingVertical: 10,
-    borderRadius: 4,
-  },
-  deleteButtonText: {
-    color: "#FFF",
-    fontSize: 16,
-  },
-  recipeIngredients: {
-    fontSize: 16,
-    marginVertical: 5,
-  },
-  recipePrice: {
-    fontSize: 16,
-    fontWeight: "bold",
-  },
-  recipeImage: {
-    height: 200,
-    marginVertical: 10,
-    borderRadius: 10,
-  },
-  totalCost: {
-    fontSize: 14,
-    marginTop: 5,
-    color: "#666",
-  },
-});
-
-SavedRecipesScreen.navigationOptions = ({ navigation }) => {
-  return {
-    headerTitle: "Saved Recs",
-    headerLeft: <MenuButton onPress={() => navigation.openDrawer()} />,
-  };
 };
 
 export default SavedRecipesScreen;
